Refactor dataflow error logic out to a class
