+++ app/gui2/src/components/ComponentBrowser/__tests__/input.test.ts
@@ -12,7 +12,7 @@ import {
   makeType,
   type SuggestionEntry,
 } from '@/stores/suggestionDatabase/entry'
-import { ComputedValueRegistry } from '@/util/computedValueRegistry'
+import { ComputedValueRegistry } from '@/stores/project/computedValueRegistry'
 import { tryIdentifier, tryQualifiedName } from '@/util/qualifiedName'
 import { unwrap } from '@/util/result'
 import type { ExprId } from 'shared/yjsModel'
