+++ app/gui2/src/stores/project/visualizationDataRegistry.ts
